fix(engine): Moving root querySelector to upgrade.ts
